move water definition into wastedata
