+++ lh_manager/waste_manager/endpoints.py
@@ -2,8 +2,9 @@
 from typing import List, Optional
 from flask import make_response, Response, request, Blueprint
 
-from ..liquid_handler.bedlayout import Well, WellLocation, Composition
-from .waste import waste_layout, WasteItem
+from ..liquid_handler.bedlayout import Well, WellLocation
+from .waste import waste_layout
+from .wastedata import WasteItem
 from .events import trigger_waste_update
 
 blueprint = Blueprint('waste_manager', __name__)
