+++ lh_manager/waste_manager/wastedata.py
@@ -1,4 +1,7 @@
-from ..liquid_handler.bedlayout import Solution
+from ..liquid_handler.bedlayout import Solution, Composition
+
+WATER = Composition(solvents=[dict(name='H2O',
+                                   fraction=1.0)])
 
 class WasteItem(Solution):
     ...
